# labels: small optimization to stringlabels